chore(suite-common): notifications and settings
