+++ packages/suite/src/config/suite/index.ts
@@ -1,5 +1,4 @@
 import { FIAT } from '@suite-common/suite-config';
 import LANGUAGES from './languages';
-import SETTINGS from './settings';
 
-export { LANGUAGES, FIAT, SETTINGS };
+export { LANGUAGES, FIAT };
