+++ suite-common/suite-config/package.json
@@ -9,6 +9,9 @@
         "lint:js": "yarn g:eslint '**/*.{ts,tsx,js}'",
         "type-check": "tsc --build"
     },
+    "dependencies": {
+        "@suite-common/suite-types": "workspace:*"
+    },
     "devDependencies": {
         "jest": "^26.6.3",
         "typescript": "4.9.5"
