+++ suite-common/suite-config/tsconfig.json
@@ -1,5 +1,5 @@
 {
     "extends": "../../tsconfig.json",
     "compilerOptions": { "outDir": "libDev" },
-    "references": []
+    "references": [{ "path": "../suite-types" }]
 }
