+++ suite-common/suite-types/src/index.ts
@@ -5,6 +5,7 @@ export * from './guide';
 export * from './sign';
 export * from './github';
 export * from './messageSystem';
+export * from './notifications';
 
 export type Selector<TReturnValue> = (state: any) => TReturnValue;
 export type SuiteCompatibleAction<TPayload> = (
