+++ suite-common/suite-types/src/notifications.ts
@@ -1,8 +1,9 @@
 import { TranslationKey } from '@suite-common/intl-types';
 import { DesktopAppUpdateState, PROTOCOL_SCHEME } from '@suite-common/suite-constants';
-import { TrezorDevice } from '@suite-common/suite-types';
 import { NetworkSymbol } from '@suite-common/wallet-config';
 import { DEVICE } from '@trezor/connect';
+
+import { TrezorDevice } from './device';
 
 export type NotificationId = number;
