+++ suite-common/toast-notifications/src/index.ts
@@ -1,6 +1,5 @@
 // This package consists only of few types now. In future we will move rest of notifications related logic here
 // because we will need that in mobile as well.
-export * from './types';
 export * from './notificationsReducer';
 export * from './notificationsActions';
 export * from './notificationsThunks';
