+++ suite-native/notifications/package.json
@@ -12,6 +12,7 @@
     "dependencies": {
         "@react-navigation/native": "^6.1.3",
         "@suite-common/icons": "workspace:*",
+        "@suite-common/suite-types": "workspace:*",
         "@suite-common/toast-notifications": "workspace:*",
         "@suite-common/wallet-config": "workspace:*",
         "@suite-common/wallet-core": "workspace:*",
