+++ suite-native/notifications/tsconfig.json
@@ -4,6 +4,9 @@
     "references": [
         { "path": "../../suite-common/icons" },
         {
+            "path": "../../suite-common/suite-types"
+        },
+        {
             "path": "../../suite-common/toast-notifications"
         },
         {
